MetaCommunityMetrics.jl

A collection of tools and utilities for analyzing meta-communities in Julia.

Welcome to the documentation for MetaCommunityMetrics. Here you can find guides and reference material on how to use the functions.

Getting Started

###Installation

To install MetaCommunityMetrics, use the following command:

using Pkg
Pkg.add("MetaCommunityMetrics")
using MetaCommunityMetrics

##Functions

MetaCommunityMetrics.beta_diversity — Method

beta_diversity(mat::Matrix; quant::Bool) -> DataFrame

Calculate beta diversity for a given biodiversity data. This function supports both binary (presence/absence) and quantitative data. For binary data, the function calculates Podani family, Jaccard-based indices. For quantitative data, the function calculates Podani family, Ruzicka-based indices. The function returns a DataFrame containing the calculated beta diversity indices. Empty patches have to be removed before calculation. Please refer to Example.jl for more details.

Arguments

mat::Matrix: A matrix where each row represents a sample and each column represents a species. The elements of the matrix should represent the presence/absence or abundance of species.

quant::Bool: A boolean flag that indicates whether the data is quantitative. Default is false, which means the data is treated as binary.

Returns

DataFrame: A DataFrame with the following columns:

BDtotal: Total beta diversity, which captures the overall dissimilarity between local communities.

Repl: Replacement component of diversity, which reflects how many species are different in one site compared to another, ignoring the species that are mere additions or subtractions.

RichDif: Richness difference component of diversity, which captures the disparity in biodiversity in terms of the count of species present, without taking into account the specific identities or distributions of those species.

MetaCommunityMetrics.mean_spatial_beta_div — Method

mean_spatial_beta_div(abundance::AbstractVector, time::AbstractVector, patch::Union{AbstractVector, String}, species::Union{AbstractVector, String}; quant::Bool) -> DataFrame

Calculate the mean spatial beta diversity components of a metacommunity over time based on species abundances or presence-absences.

Arguments

abundance::Vector: A vector containing abundance data for each species across different samples.

time::Vector: A vector indicating the time each sample was taken.

patch::Vector: A vector indicating the spatial location (patch) of each sample.

species::Vector: A vector indicating the species associated with each abundance entry.

quant::Bool: Optional boolean flag to indicate whether the data should be treated as quantitative (default is false, treating data as binary presence/absence).

Returns

DataFrame: A DataFrame containing the mean values of total beta diversity, replacement, and richness difference components across all time points. Columns are mean_spatial_BDtotal, mean_spatial_Repl, and mean_spatial_RichDif.

MetaCommunityMetrics.mean_temporal_beta_div — Method

mean_temporal_beta_div(abundance::AbstractVector, time::AbstractVector, patch::Union{AbstractVector, String}, species::Union{AbstractVector, String};quant::Bool) -> DataFrame

Calculate the mean temporal beta diversity components acorss all patches based on species abundances or presence-absences.

Arguments

abundance::Vector: A vector containing abundance data for each species across different samples.

time::Vector: A vector indicating the time each sample was taken.

patch::Vector: A vector indicating the spatial location (patch) of each sample.

species::Vector: A vector indicating the species associated with each abundance entry.

quant::Bool: Optional boolean flag to indicate whether the data should be treated as quantitative (default is false, treating data as binary presence/absence).

Returns

DataFrame: A DataFrame containing the mean values of total beta diversity, replacement, and richness difference components across all pactches. Columns are mean_temporal_BDtotal, mean_temporal_Repl, and mean_temporal_RichDif.

MetaCommunityMetrics.create_clusters — Method

create_clusters(time::Vector{Int}, latitude::Vector{Float64}, longitude::Vector{Float64}, patch::Vector{Int}) -> Dict{Int, DataFrame}

Create clusters for each unique time step in a dataset. Only presnece-absence data can be used.

Arguments

time::Vector: A vector indicating the time each sample was taken.

latitude::Vector: A vector indicating the latitude of each sample.

longitude::Vector: A vector indicating the longitude of each sample.

patch::Vector: A vector indicating the spatial location (patch) of each sample.

Returns

Dict{Int, DataFrame}: A dictionary where each key is a unique time from the dataset and each value is a DataFrame for that time with an added Group column indicating the assigned cluster.

Details

This function performs hierarchical clustering on geographical coordinates for each unique time step. It aims to balance the clusters based on the number of sites and species richness, ensuring that no group has less than five sites and that there are at least two groups. If conditions for clustering balance are not met (like groups having less than five sites or only one group), it iteratively adjusts the clusters by reassigning sites to improve group balance.

Example

using CSV, DataFrames
using DataFramesMeta
using Pipe: @pipe

sample_df = @pipe CSV.read("data/rodent_abundance_data.csv", DataFrame; header=true) |>#read in the sample data
            select(_, Not(:Column1))|> #select the columns 
            stack(_, Not(:Sampling_date_order, :Year, :Month, :Day, :plot), variable_name = :species, value_name = :abundance)

preped_data= @pipe sample_df |>         
@transform(_, :presence = ifelse.(:abundance .>= 1, 1, 0)) |>
groupby(_, [:Sampling_date_order, :species]) |> 
combine(_,:presence=>sum=>:total_presence) |>
filter(row -> row[:total_presence] !<= 1, _) |> #remove singletons (species occurring at one site only)
leftjoin(_,sample_df, on = [:Sampling_date_order, :species]) |> #join the data back to the original data

# Generate random latitude and longitude values
n = nrow(sample_df)
latitude = rand(35.0:0.01:36.0, n) # Adjust range as needed
longitude = rand(-120.0:0.01:-119.0, n) # Adjust range as needed

# Add the coordinates to the DataFrame
sample_df[:, :latitude] = latitude
sample_df[:, :longitude] = longitude

# Create groups for each time step
grouped_data = create_clusters(sample_df.Sampling_date_order, sample_df.latitude, sample_df.longitude, sample_df.plot)

MetaCommunityMetrics.plot_clusters — Method

plot_clusters(grouped_data::DataFrame)

This function plots the clustering result from the function create_clusters() on a scatter plot according to geographic coordinates. Each point represents a location, and points are colored based on their group/cluster ID.

Arguments

grouped_data::DataFrame: A DataFrame containing the columns Latitude, Longitude, and Group.

Latitude: A column of latitude coordinates.

Longitude: A column of longitude coordinates.

Group: A column indicating the cluster ID for each point.

Returns

Returns

This function displays a scatter plot of the clusters.

Details

The function extracts the Latitude, Longitude, and Group columns from the provided DataFrame.

Unique cluster IDs are assigned numeric identifiers.

A color palette is generated using distinguishable colors to differentiate the clusters.

The scatter plot is created with points colored by their cluster ID.

Example

using DataFrames
# Create a sample DataFrame
data = DataFrame(Latitude = [34.05, 36.16, 40.71, 34.05],
                 Longitude = [-118.24, -115.15, -74.01, -118.24],
                 Group = ["A", "B", "A", "B"])

# Plot clusters
plot_clusters(data)

MetaCommunityMetrics.DNCI_multigroup — Function

DNCI_multigroup(comm::Matrix, groups::Vector, Nperm::Int=1000, count::Bool=true)

Calculates the dispersal-niche continuum index (DNCI) for multiple groups, a metric proposed by Vilmi(2021) (doi: 10.1111/ecog.05356).

Arguments

comm::Matrix: A presence-absence data matrix where rows represent observations (e.g., sites or samples) and columns represent species.

groups::Vector: A vector indicating the group membership for each row in the comm matrix.

Nperm::Int=1000: The number of permutations for significance testing. Default is 1000.

count::Bool=true: A flag indicating whether the numeber of permutations is printed. Default is false.

Returns

DataFrame: A DataFrame containing the DNCI results for each pair of groups.

Examples

```julia

Example usage of DNCI_multigroup

comm = [1 0 0 1 0; 1 1 0 0 0; 0 1 1 0 0; 0 0 1 1 1; 1 0 0 0 1; 0 1 1 0 1]

groups = ["A", "A", "B", "B", "C", "C"] Nperm = 1000 count = true

result = DNCI_multigroup(comm, groups, Nperm, count) println(result)