Beta Diversity Functions

The beta diversity functions in MetaCommunityMetrics are adapted from the beta.div.comp function in the R package adespatial. These methods were originally developed and detailed by Pierre Legendre(2014). The implementation in Julia provides a more efficient means to compute these metrics, particularly for large-scale datasets, while maintaining the robustness of the original methodology.

Beta diversity is a fundamental concept in ecology that quantifies the variation in species composition between different habitats, plots, or over time. In the context of metacommunity analysis, beta diversity functions help to assess how community composition changes spatially across different locations or temporally within a given location.

Choosing the Right Function

Use beta_diversity when you want a general overview of diversity across your dataset.

Opt for mean_spatial_beta_div when your focus is on comparing diversity between different spatial locations.

Select mean_temporal_beta_div to track how diversity changes over time within the same location.

The Functions

MetaCommunityMetrics.beta_diversity — Function

beta_diversity(mat::Matrix; quant::Bool) -> DataFrame

Calculate beta diversity for a given biodiversity data. This function supports both binary (presence/absence) and quantitative data. For binary data, the function calculates Podani family, Jaccard-based indices. For quantitative data, the function calculates Podani family, Ruzicka-based indices. The function returns a DataFrame containing the calculated beta diversity indices. Empty patches have to be removed before calculation.

Arguments

mat::Matrix: A matrix where each row represents a sample and each column represents a species. The elements of the matrix should represent the presence/absence or abundance of species.

quant::Bool: A boolean flag that indicates whether the data is quantitative. Default is false, which means the data is treated as binary.

Returns

DataFrame: A DataFrame with the following columns:

BDtotal: Total beta diversity, which captures the overall dissimilarity between local communities.

Repl: Replacement component of diversity, which reflects how many species are different in one site compared to another, ignoring the species that are mere additions or subtractions.

RichDif: Richness difference component of diversity, which captures the disparity in biodiversity in terms of the count of species present, without taking into account the specific identities or distributions of those species.

MetaCommunityMetrics.mean_spatial_beta_div — Function

mean_spatial_beta_div(abundance::AbstractVector, time::AbstractVector, patch::Union{AbstractVector, String}, species::Union{AbstractVector, String}; quant::Bool) -> DataFrame

Calculate the mean spatial beta diversity components of a metacommunity over time based on species abundances or presence-absences.

Arguments

abundance::Vector: A vector containing abundance data for each species across different samples.

time::Vector: A vector indicating the time each sample was taken.

patch::Vector: A vector indicating the spatial location (patch) of each sample.

species::Vector: A vector indicating the species associated with each abundance entry.

quant::Bool: Optional boolean flag to indicate whether the data should be treated as quantitative (default is false, treating data as binary presence/absence).

Returns

DataFrame: A DataFrame containing the mean values of total beta diversity, replacement, and richness difference components across all time points. Columns are mean_spatial_BDtotal, mean_spatial_Repl, and mean_spatial_RichDif.

MetaCommunityMetrics.mean_temporal_beta_div — Function

mean_temporal_beta_div(abundance::AbstractVector, time::AbstractVector, patch::Union{AbstractVector, String}, species::Union{AbstractVector, String};quant::Bool) -> DataFrame

Calculate the mean temporal beta diversity components acorss all patches based on species abundances or presence-absences.

Arguments

abundance::Vector: A vector containing abundance data for each species across different samples.

time::Vector: A vector indicating the time each sample was taken.

patch::Vector: A vector indicating the spatial location (patch) of each sample.

species::Vector: A vector indicating the species associated with each abundance entry.

quant::Bool: Optional boolean flag to indicate whether the data should be treated as quantitative (default is false, treating data as binary presence/absence).

Returns

DataFrame: A DataFrame containing the mean values of total beta diversity, replacement, and richness difference components across all pactches. Columns are mean_temporal_BDtotal, mean_temporal_Repl, and mean_temporal_RichDif.