Dispersal-Niche Continuum Index (DNCI) Functions

The Dispersal-Niche Continuum Index (DNCI) functions in MetaCommunityMetrics are adapted from the R package DNCImper: Assembly process identification based on SIMPER analysis. These methods, originally developed by Clarke in 1993 and later refined by Gibert & Escarguel in 2019 and Vilmi, Gibert et al. in 2021, offer powerful tools for identifying the processes underlying species assembly in metacommunities.

Background

The DNCI functions is built around the Per-SIMPER and DNCI analyses. PerSIMPER, based on the Similarity Percentage (SIMPER) analysis developed by Clarke (1993), assesses the contribution of individual taxa to overall dissimilarity (OAD) between groups of assemblages. PerSIMPER enhances this by comparing empirical SIMPER plots with randomized plots generated through matrix permutation, which helps identify whether niche, dispersal, or both processes are driving community assembly.

The DNCI (Dispersal-Niche Continuum Index) further extends this approach by transforming the qualitative results of PerSIMPER into a quantitative index, providing a straightforward measure of the influence of niche and dispersal processes on community structure.

Functionality Overview

The DNCI functions in MetaCommunityMetrics allow you to analyze the processes driving species assembly within your dataset. By comparing empirical data with randomized permutations, you can determine the extent to which niche and dispersal processes have influenced the structure of metacommunities.

The Functions

create_clusters: Groups sampling locations based on their spatial attributes and species richness, which can then be used to assess DNCI.

plot_clusters: Visualizes the clusters created, allowing for an intuitive understanding of spatial groupings.

DNCI_multigroup: Computes the Dispersal-Niche Continuum Index (DNCI) across multiple groups, helping to quantify the relative influence of niche versus dispersal processes.

MetaCommunityMetrics.create_clusters — Function

create_clusters(time::Vector{Int}, latitude::Vector{Float64}, longitude::Vector{Float64}, patch::Vector{Int}, total_richness::Vector{Int}) -> Dict{Int, DataFrame}

Create clusters for each unique time step in a dataset. Only presnece-absence data can be used.

Arguments

time::Vector: A vector indicating the time each sample was taken.

latitude::Vector: A vector indicating the latitude of each sample.

longitude::Vector: A vector indicating the longitude of each sample.

patch::Vector: A vector indicating the spatial location (patch) of each sample.

total_richness::Vector: A vector indicating the total species richness at each plot at each time step.

Returns

Dict{Int, DataFrame}: A dictionary where each key is a unique time from the dataset and each value is a DataFrame for that time with an added Group column indicating the assigned cluster.

Details

This function performs hierarchical clustering on geographical coordinates for each unique time step. It aims to balance the clusters based on the number of sites and species richness, ensuring that no group has less than five sites and that there are at least two groups. If conditions for clustering balance are not met (like groups having less than five sites or only one group), it iteratively adjusts the clusters by reassigning sites to improve group balance.

MetaCommunityMetrics.plot_clusters — Function

plot_clusters(latitude::Vector{Float64}, longitude::Vector{Float64}, group::Union{AbstractVector, String})

Plots clusters of data points on a scatter plot using their geographic coordinates and cluster assignments.

Arguments

latitude::Vector{Float64}: A vector of latitude coordinates.

longitude::Vector{Float64}: A vector of longitude coordinates.

group::Union{AbstractVector, String}: A vector or string indicating the cluster assignments for each data point.

Details

The function assigns a unique color to each cluster and plots the points based on their geographic coordinates.

The points are colored according to their cluster assignment.

The plot includes black borders around the markers for better visibility.

Example

latitudes = [35.0, 35.0, 35.5, 35.5, 35.5, 36.0, 36.0]
longitudes = [-110.0, -109.5, -109.5, -109.0, -108.0, -109.5, -108.0]
groups = [1, 1, 1, 1, 2, 1, 2]

plot_clusters(latitudes, longitudes, groups)

MetaCommunityMetrics.DNCI_multigroup — Function

DNCI_multigroup(comm::Matrix, groups::Vector, Nperm::Int=1000, count::Bool=true) -> DataFrame

Calculates the dispersal-niche continuum index (DNCI) for multiple groups, a metric proposed by Vilmi(2021) (doi: 10.1111/ecog.05356).

Arguments

comm::Matrix: A presence-absence data matrix where rows represent observations (e.g., sites or samples) and columns represent species.

groups::Vector: A vector indicating the group membership for each row in the comm matrix.

Nperm::Int=1000: The number of permutations for significance testing. Default is 1000.

count::Bool=true: A flag indicating whether the numeber of permutations is printed. Default is false.

Returns

DataFrame: A DataFrame containing the DNCI results for each pair of groups.

Examples

# Example usage of DNCI_multigroup
comm = [1 0 0 1 0;
        1 1 0 0 0;
        0 1 1 0 0;
        0 0 1 1 1;
        1 0 0 0 1;
        0 1 1 0 1]

groups = ["A", "A", "B", "B", "C", "C"]
Nperm = 1000
count = true

result = DNCI_multigroup(comm, groups, Nperm, count)
println(result)