Hypervolume Functions

Hypervolume is a method originally proposed by Hutchinson (1957). It provide methods to calculate the volume of a niche for a given specice and thus the ncihe overlap between two species. It helps to infer how niche breath of species has contribute to the co-occurance of different species in the same location at the same time. The hypervolume functions provide by this package are adapted from the R package MVNH (https://github.com/lvmuyang/MVNH)

The Functions

MetaCommunityMetrics.MVNH_det — Function

MVNH_det(data::DataFrame; var_names::Vector{String}=String[]) -> DataFrame

Calculate the niche volume for a species in a given biodiversity data.

Arguments

data::DataFrame: A dataframe where each row represents an obsevation of species and each column represent the environmental variables for that observation.

var_names::Vector{String}: A vector of strings with the names of the environmental variables. Default is an empty vector.

Returns

DataFrame: A DataFrame with the following columns:

total: The total hypervolume of the niche.

cor: The correlation component of the hypervolume, also known as the determinant of the correlation matrix.

variable_i: The variance of the environmental variable i.

Details

Evirnomental variables are assumed to be normally distributed.

Evirnomental variables are encouraged to be normalized to avoid bias due to different scaling before using this function.

This function is a translation/adaptation of the MVNH_det function from the R package MVNH,licensed under GPL-3.

Original package and documentation available at: https://github.com/lvmuyang/MVNH

Example

julia> using MetaCommunityMetrics, Pipe, DataFrames, Statistics

julia> df = load_sample_data()
48735×10 DataFrame
   Row │ Year   Month  Day    Sampling_date_order  plot   Species  Abundance  Presence  Latitude  Longitude 
       │ Int64  Int64  Int64  Int64                Int64  String3  Int64      Int64     Float64   Float64   
───────┼────────────────────────────────────────────────────────────────────────────────────────────────────
     1 │  2010      1     16                    1      1  BA               0         0      35.0     -110.0
     2 │  2010      1     16                    1      2  BA               0         0      35.0     -109.5
     3 │  2010      1     16                    1      8  BA               0         0      35.5     -109.5
     4 │  2010      1     16                    1      9  BA               0         0      35.5     -109.0
     5 │  2010      1     16                    1     11  BA               0         0      35.5     -108.0
   ⋮   │   ⋮      ⋮      ⋮             ⋮             ⋮       ⋮         ⋮         ⋮         ⋮          ⋮
 48731 │  2023      3     21                  117      9  SH               0         0      35.5     -109.0
 48732 │  2023      3     21                  117     10  SH               0         0      35.5     -108.5
 48733 │  2023      3     21                  117     12  SH               1         1      35.5     -107.5
 48734 │  2023      3     21                  117     16  SH               0         0      36.0     -108.5
 48735 │  2023      3     21                  117     23  SH               0         0      36.5     -108.0
                                                                                          48725 rows omitted

julia> result = MVNH_det(df; var_names=["Latitude", "Longitude"])

MetaCommunityMetrics.MVNH_dissimilarity — Function

MVNH_dissimilarity(data_1::DataFrame, data_2::DataFrame; var_names::Vector{String}=String[]) -> Dict

Calculate the niche overlap between two species in a given biodiversity data.

MetaCommunityMetrics.average_MVNH_det — Function

average_MVNH_det(data::DataFrame, presence_absence::Vector{Int}, species::Vector{String}; var_names::Vector{String}=String[]) -> Number

Calculate the avaraged niche volume across all species in a given biodiversity data.

Returns

Number: The computed average niche volume based on the MVNH_det functions.

MetaCommunityMetrics.average_MVNH_dissimilarity — Function

MVNH_dissimilarity(data_1::DataFrame, data_2::DataFrame; var_names::Vector{String}=String[]) -> Number

Calculate the avearaged niche overlap across all species pairs in a given biodiversity data.

Returns

Number: The computed average niche overlap based on the MVNH_dissimilarity functions.