Variability Metrics Functions

The variability metrics functions in MetaCommunityMetrics are designed to capture changes in dispersal and density-dependent biotic interactions by investigating temporal variability and synchrony across spatial scales and organizational levels within a metacommunity. These functions are based on the work of Wang et al. (2019), which provides a framework for quantifying variability at different scales and contexts within a community.

Overview

In MetaCommunityMetrics, the CV_meta_simple function is directly adapted from the R function var.partition in Wang et al. (2019). This function is designed with computational efficiency in mind, particularly for large datasets, by avoiding the calculation of all covariances between species. This approach ensures faster performance while still providing valuable insights into variability across different scales within a metacommunity.

In contrast, the CV_meta function extends the analysis by including the calculation of all covariances between species, offering a more detailed and comprehensive examination of variability. This approach captures interactions between species that may be overlooked by more streamlined methods. While the CV_meta function provides a richer analysis, the CV_meta_simple function remains a valuable tool when computational efficiency is a priority.

These metrics are designed to quantify variability at different scales and contexts within the community:

Local-scale average species variability (CV_s_l)

Regional-scale average species variability (CV_s_r)

Local-scale average community variability (CV_c_l)

Regional-scale community variability (CV_c_r)

The variability metrics are calculated as follows:

The variability metrics CV_s_l, CV_s_r, CV_c_l, and CV_c_r are set to zero whenever the mean abundance equals zero at any combination of spatial scales (a patch/all patches) and species number (a species/the whole community). This approach allows us to assess the impact of spatial scale on variability and to understand how different factors influence community dynamics across scales.

The Function

MetaCommunityMetrics.CV_meta — Function

CV_meta(abundance::AbstractVector, time::AbstractVector, patch::Union{AbstractVector, String}, species::Union{AbstractVector, String}) -> DataFrame

Calculates various coefficients of variation (CV) for species and community biomass at both local and regional scales within a metacommunity.

Arguments

abundance::AbstractVector: A vector representing the abundance of species.

time::AbstractVector: A vector representing the time points at which the abundance measurements were taken.

patch::Union{AbstractVector, String}: A vector or single value representing the patch or plot identifier.

species::Union{AbstractVector, String}: A vector or single value representing the species identifier.

Returns

DataFrame: A DataFrame containing the following columns:

CV_s_l: Local-scale average species variability.

CV_s_r: Regional-scale average species variability.

CV_c_l: Local-scale average community variability.

CV_c_r: Regional-scale community variability.

Details This function calculates the coefficients of variation (CV) for species and community biomass at both local and regional scales. The calculation involves several steps:

Reorganization of Data: The input data is organized into a DataFrame with columns for abundance, time, plot, and species.

Mean Calculations: Temporal mean species abundance is calculated for each species in each patch, as well as the overall temporal mean biomass.

Temporal Variance Calculations: Temporal variance is calculated for each species within patches, for species across patches, for the community biomass within patches, and for the overall metacommunity biomass.

CV Calculations: The coefficients of variation are calculated for species and community biomass at both local and regional scales.

Output: The results are returned in a DataFrame summarizing the CVs for local and regional scales.

Example

julia> using MetaCommunityMetrics, Pipe

julia> df = @pipe load_sample_data() |>
                          filter(row -> row[:Sampling_date_order] < 20, _)
6764×10 DataFrame
  Row │ Year   Month  Day    Sampling_date_order  plot   Species  Abundance  Presence  Latitude  Longitude 
      │ Int64  Int64  Int64  Int64                Int64  String3  Int64      Int64     Float64   Float64   
──────┼────────────────────────────────────────────────────────────────────────────────────────────────────
    1 │  2010      1     16                    1      1  BA               0         0      35.0     -110.0
    2 │  2010      1     16                    1      2  BA               0         0      35.0     -109.5
    3 │  2010      1     16                    1      8  BA               0         0      35.5     -109.5
    4 │  2010      1     16                    1      9  BA               0         0      35.5     -109.0
    5 │  2010      1     16                    1     11  BA               0         0      35.5     -108.0
  ⋮   │   ⋮      ⋮      ⋮             ⋮             ⋮       ⋮         ⋮         ⋮         ⋮          ⋮
 6760 │  2012      1     21                   19     21  SH               0         0      36.5     -109.0
 6761 │  2012      1     21                   19      5  SH               0         0      35.0     -108.0
 6762 │  2012      1     21                   19      7  SH               0         0      35.5     -110.0
 6763 │  2012      1     21                   19     23  SH               0         0      36.5     -108.0
 6764 │  2012      1     21                   19     24  SH               0         0      36.5     -107.5
                                                                                          6754 rows omitted
                                                                                          
julia> CV_summary_df = CV_meta(df.Abundance, df.Sampling_date_order, df.plot, df.Species)
1×4 DataFrame
 Row │ CV_s_l   CV_s_r    CV_c_l    CV_c_r   
     │ Float64  Float64   Float64   Float64  
─────┼───────────────────────────────────────
   1 │ 1.05607  0.813324  0.656184  0.537299

MetaCommunityMetrics.CV_meta_simple — Function

CV_meta_simple(abundance::AbstractVector, time::AbstractVector, patch::Union{AbstractVector, String}, species::Union{AbstractVector, String}) -> DataFrame

Calculates coefficients of variation (CV) for species and community biomass at both local and regional scales within a metacommunity, using a simpler approach optimized for handling larger datasets.

Arguments

abundance::AbstractVector: A vector representing the abundance of species.

time::AbstractVector: A vector representing the time points at which the abundance measurements were taken.

patch::Union{AbstractVector, String}: A vector or single value representing the patch or plot identifier.

species::Union{AbstractVector, String}: A vector or single value representing the species identifier.

Returns

DataFrame: A DataFrame containing the following columns:

CV_s_l: Local-scale average species variability.

CV_s_r: Regional-scale average species variability.

CV_c_l: Local-scale average community variability.

CV_c_r: Regional-scale community variability.

Details This function is a simplified version of the CV_meta function, designed to efficiently handle larger datasets by avoiding complex covariance calculations. The steps include:

Reorganization of Data: The input data is organized into a DataFrame with columns for abundance, time, plot, and species, and then transformed into a 3D abundance matrix.

Total Abundance Calculations: The function calculates total abundances for species across time, within each patch, and for the entire metacommunity.

Standard Deviation (SD) Calculations: Temporal standard deviations of abundance are computed for the entire metacommunity, each patch, and each species.

CV Calculations: The coefficients of variation are calculated for species and community biomass at both local and regional scales.

Output: The results are returned in a DataFrame summarizing the CVs for local and regional scales.

Example

julia> using MetaCommunityMetrics, Pipe

julia> df = @pipe load_sample_data() |>
                          filter(row -> row[:Sampling_date_order] < 20, _)
6764×10 DataFrame
  Row │ Year   Month  Day    Sampling_date_order  plot   Species  Abundance  Presence  Latitude  Longitude 
      │ Int64  Int64  Int64  Int64                Int64  String3  Int64      Int64     Float64   Float64   
──────┼────────────────────────────────────────────────────────────────────────────────────────────────────
    1 │  2010      1     16                    1      1  BA               0         0      35.0     -110.0
    2 │  2010      1     16                    1      2  BA               0         0      35.0     -109.5
    3 │  2010      1     16                    1      8  BA               0         0      35.5     -109.5
    4 │  2010      1     16                    1      9  BA               0         0      35.5     -109.0
    5 │  2010      1     16                    1     11  BA               0         0      35.5     -108.0
  ⋮   │   ⋮      ⋮      ⋮             ⋮             ⋮       ⋮         ⋮         ⋮         ⋮          ⋮
 6760 │  2012      1     21                   19     21  SH               0         0      36.5     -109.0
 6761 │  2012      1     21                   19      5  SH               0         0      35.0     -108.0
 6762 │  2012      1     21                   19      7  SH               0         0      35.5     -110.0
 6763 │  2012      1     21                   19     23  SH               0         0      36.5     -108.0
 6764 │  2012      1     21                   19     24  SH               0         0      36.5     -107.5
                                                                                          6754 rows omitted

julia> CV_summary_df = CV_meta_simple(df.Abundance, df.Sampling_date_order, df.plot, df.Species)
1×4 DataFrame
 Row │ CV_s_l   CV_s_r    CV_c_l    CV_c_r   
     │ Float64  Float64   Float64   Float64  
─────┼───────────────────────────────────────
   1 │ 1.20808  0.918862  0.801216  0.665082

References

Wang, S., Lamy, T., Hallett, L. M. & Loreau, M. Stability and synchrony across ecological hierarchies in heterogeneous metacommunities: linking theory to data. Ecography 42, 1200-1211 (2019). https://doi.org:https://doi.org/10.1111/ecog.04290